requirements.txt
datarobot==3.0.2
numpy==1.23.3
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
streamlit==1.19.0
openai==0.28
shap
boto3
xgboost
shap
streamlit_shap
